chore: Move class to separate file
